Refactor: Extract submission photo upload flow to PhotoUploader
- Extracted `uploadSubmitPhotos` from `UploadManager` into a new dedicated `PhotoUploader` class.
- Modified `UploadManager` to delegate to `PhotoUploader`, reducing its responsibilities.
- Updated `ServiceModule` to inject the new dependency and pass it to `UploadManager`.
- Updated `UploadManagerTest` to maintain full test coverage for the delegated behavior.
