+++ app/src/test/java/org/ole/planet/myplanet/services/UploadManagerTest.kt
@@ -40,7 +40,7 @@ import org.ole.planet.myplanet.repository.VoicesRepository
 import org.ole.planet.myplanet.services.upload.UploadConfigs
 import org.ole.planet.myplanet.services.upload.UploadCoordinator
 import org.ole.planet.myplanet.services.upload.UploadResult
-import org.ole.planet.myplanet.services.upload.delegate.PhotoUploadDelegate
+import org.ole.planet.myplanet.services.upload.PhotoUploader
 import org.ole.planet.myplanet.utils.TestDispatcherProvider
 
 @OptIn(ExperimentalCoroutinesApi::class)
@@ -61,7 +61,7 @@ class UploadManagerTest {
     private val teamsRepository: Lazy<TeamsRepository> = mockk(relaxed = true)
     private val apiInterface: ApiInterface = mockk(relaxed = true)
     private val activitiesRepository: ActivitiesRepository = mockk(relaxed = true)
-    private lateinit var photoUploadDelegate: PhotoUploadDelegate
+    private lateinit var photoUploader: PhotoUploader
 
     private val testDispatcher = StandardTestDispatcher()
     private val testScope = TestScope(testDispatcher)
@@ -76,7 +76,7 @@ class UploadManagerTest {
         every { Log.e(any(), any()) } returns 0
         every { Log.e(any(), any(), any()) } returns 0
 
-        photoUploadDelegate = PhotoUploadDelegate(submissionsRepository, apiInterface, TestDispatcherProvider(testDispatcher))
+        photoUploader = PhotoUploader(submissionsRepository, apiInterface, TestDispatcherProvider(testDispatcher))
 
         uploadManager = spyk(
             UploadManager(
@@ -97,7 +97,7 @@ class UploadManagerTest {
                 activitiesRepository,
                 TestDispatcherProvider(testDispatcher),
                 testScope,
-                photoUploadDelegate
+                photoUploader
             )
         )
     }
